+++ geb/model.py
@@ -550,7 +550,7 @@ class GEBModel(Module):
         self.reporter.finalize()
         self.create_done_file()
 
-    def run_yearly(self) -> None:
+    def run_yearly(self, model_name: str = "default") -> None:
         """Run the model in yearly mode, where timesteps are yearly rather than daily.
 
         This depends on a spinup run that was run in daily mode.
@@ -558,7 +558,8 @@ class GEBModel(Module):
         Notes:
             Cannot be run in combination with hydrology simulation.
             This mode is experimential and is not fully tested.
-
+        Args:
+            model_name: Name of the model run. This is used to create a folder for the results of the model run. Defaults to "default".
         Raises:
             ValueError: If the start or end time is not at the beginning or end of a year, respectively.
             ValueError: If flood simulation is enabled in the config, as this is not compatible with yearly mode.
@@ -571,6 +572,8 @@ class GEBModel(Module):
             for key, value in self.config["report"].items()
             if key.startswith("agents.households")
         }
+
+        self.config["general"]["name"] = model_name
 
         if self.config["hazards"]["floods"]["simulate"] is True:
             raise ValueError(